fix(flutter): make the Dart side compile against FRB 2.12.0 bindings
The web release build failed in `dart compile js` because the hand-written
Dart did not match what flutter_rust_bridge 2.12.0 actually generates:

- bridge_rust.dart imported `src/rust/api/mod.dart`, but FRB mirrors the
  Rust module `crate::api` (a mod.rs) onto `src/rust/api.dart`. The missing
  import made every `ffi.*` symbol undefined. Fix the import path.
- FRB maps Rust `u64` -> Dart `BigInt`, so the engine handles cross the
  boundary as BigInt while the app talks in `int`. Convert int<->BigInt in
  the adapter (the one place that knows the generated types): `_h()` for
  handle args, `?.toInt()` for `create_*` results.
- `pointer`'s `button: i64` mapped to `PlatformInt64` (BigInt on web), which
  the `int` call site couldn't satisfy. Narrow it to `i32` (the engine
  clamps button to u8 anyway) so it maps to a plain Dart `int`.
- elpa_shell.dart used PointerSignalEvent/PointerScrollEvent without
  importing them; widgets.dart's re-export `show` list omits both. Add an
  explicit `package:flutter/gestures.dart` import.

Also pin `flutter_rust_bridge = "=2.12.0"` to match the codegen version
(as the README requires); codegen applies this pin automatically.

+++ flutter/rust/src/api/mod.rs
@@ -152,8 +152,8 @@ pub fn dispose(handle: u64) -> bool {
 
 /// Forward a pointer event (logical coordinates); returns emitted UI messages.
 #[frb(sync)]
-pub fn pointer(handle: u64, phase: FfiPointer, x: f64, y: f64, button: i64) -> Vec<FfiMessage> {
-    with(handle, |e| into_ffi(e.pointer(phase.into(), x, y, button))).unwrap_or_default()
+pub fn pointer(handle: u64, phase: FfiPointer, x: f64, y: f64, button: i32) -> Vec<FfiMessage> {
+    with(handle, |e| into_ffi(e.pointer(phase.into(), x, y, button as i64))).unwrap_or_default()
 }
 
 /// Forward a scroll/wheel event.
